Pre-validation (#17)
* Refactors & lifts form table state

* Changes cell style via play validation function
