+++ webapp/src/main/new/New.tsx
@@ -1,16 +1,18 @@
-import { useState } from "react";
+import { useState, useMemo } from "react";
 
 import { QuadratDataEntryForm } from "./QuadratDataEntryForm";
 import { QuadratMetadataEntryForm } from "./QuadratMetadataEntryForm";
 import { Stem } from "../../types";
 import { insertStems } from "./dataService";
+import { columns } from "./QuadratDataEntryForm/columnHeaders";
+import { mockCensusData } from "../../mockData/mockCensusData";
 
 export const New = () => {
   const [errors, setErrors] = useState<any>([]);
 
   const hardCodedFormData: Stem[] = [
     {
-      Subquadrat: "11",
+      Subquadrat: 11,
       Tag: 1,
       SpCode: "species",
       DBH: 10,
@@ -19,7 +21,7 @@ export const New = () => {
       Comments: "",
     },
     {
-      Subquadrat: "11",
+      Subquadrat: 11,
       Tag: 2,
       SpCode: "species",
       DBH: 10,
@@ -30,11 +32,14 @@ export const New = () => {
   ];
 
   const [formData] = useState(hardCodedFormData);
+  const columnHeaders = useMemo(() => columns, []);
+  const data = useMemo(() => mockCensusData, []);
+
   return (
     <>
-      <h1>New Plant Form</h1>
+      <h1>Old Trees Form</h1>
       <QuadratMetadataEntryForm />
-      <QuadratDataEntryForm />
+      <QuadratDataEntryForm columns={columnHeaders} initialData={data} />
       <button
         type="submit"
         onClick={() => {
