+++ webapp/src/main/new/QuadratDataEntryForm/columnHeaders.ts
@@ -1,0 +1,30 @@
+export const columns = [
+  {
+    Header: "Tag",
+    accessor: "Tag",
+  },
+  {
+    Header: "Subquadrat",
+    accessor: "Subquadrat",
+  },
+  {
+    Header: "SpCode",
+    accessor: "SpCode",
+  },
+  {
+    Header: "DBH",
+    accessor: "DBH",
+  },
+  {
+    Header: "Htmeas",
+    accessor: "Htmeas",
+  },
+  {
+    Header: "Codes",
+    accessor: "Codes",
+  },
+  {
+    Header: "Comments",
+    accessor: "Comments",
+  },
+]
